requirements.txt
alembic==1.0.3
attrs==18.2.0
Babel==2.6.0
BTrees==4.5.1
certifi==2018.10.15
chardet==3.0.4
clld==4.4.2
clldutils==2.5.2
colander==1.5.1
colorlog==3.1.4
configparser==3.5.0
csvw==1.4.3
et-xmlfile==1.0.1
ExtensionClass==4.4
feedparser==5.2.1
-e git+git@github.com:oopcode/gld.git@17128412717f11020d9f4fb6b319c5d46e28d6d0#egg=gld
hupper==1.4.1
idna==2.7
iso8601==0.1.12
isodate==0.6.0
jdcal==1.4
latexcodec==1.0.5
Mako==1.0.7
MarkupSafe==1.1.0
nameparser==1.0.2
openpyxl==2.5.11
paginate==0.5.6
PasteDeploy==1.5.2
persistent==4.4.3
plaster==1.0
plaster-pastedeploy==0.6
psycopg2-binary==2.7.6.1
purl==1.4
pybtex==0.22.0
pycldf==1.5.2
pyparsing==2.3.0
pyramid==1.10.1
pyramid-mako==1.0.2
pyramid-tm==2.2.1
python-dateutil==2.7.5
python-editor==1.0.3
python-gettext==3.0
pytz==2018.7
PyYAML==3.13
rdflib==4.2.2
requests==2.20.1
rfc3986==1.1.0
six==1.11.0
SQLAlchemy==1.2.14
tabulate==0.8.2
transaction==2.4.0
translationstring==1.3
uritemplate==3.0.0
urllib3==1.24.1
venusian==1.1.0
waitress==1.1.0
webassets==0.12.1
WebHelpers2==2.0
WebOb==1.8.4
xlwt==1.3.0
yuicompressor==2.4.8
zc.lockfile==1.4
ZConfig==3.3.0
zdaemon==4.3
ZEO==5.2.0
ZODB==5.5.1
ZODB3==3.11.0
zodbpickle==1.0.2
zope.browser==2.3
zope.browsermenu==4.4
zope.browserresource==4.3
zope.component==4.5
zope.configuration==4.3.0
zope.contenttype==4.4
zope.deferredimport==4.3
zope.deprecation==4.3.0
zope.event==4.4
zope.exceptions==4.3
zope.hookable==4.2.0
zope.i18n==4.6.1
zope.i18nmessageid==4.3.1
zope.interface==4.6.0
zope.location==4.2
zope.pagetemplate==4.4.1
zope.proxy==4.3.1
zope.publisher==5.0.1
zope.schema==4.9.3
zope.security==4.3.0
zope.sqlalchemy==1.0
zope.tal==4.4
zope.tales==4.3
zope.traversing==4.3.1
